Add the rest of the voice impl
This is mostly the rest of the stuff that I hadn't committed yet.  

Adds:
voice_client function additions and changes (listen, stop, etc)
Gateway event and ssrc handling
The AudioReader and sink objects
The OpusRouter in its current state

I still need to do another pass on most of this to clean and fix all the messy bits but this is basically the rest of it.